Fix max density grid computation for NanoVDB volumes.
The previous version was buggy in that by looping over leaves, it didn't
consider active voxels at higher levels of the tree.

It also seems to have been hurting performance by using the leaf's maximum
value to set the density. Considering density pointwise in the nanovdb grid
ends up giving an overall 1.37x speedup for bunny-cloud and 2.80x for
explosion(!!)