Fix subtle bug with dividing by the wavelength PDFs.
Previously, we'd add up contributions for all the vertices of a path and
then divide by wavengths' PDFs in Film::AddSample()/AddSplat(). The problem
with this was the case where secondary wavelengths are terminated in the
middle of path (e.g. for dispersion). What we actually should be doing (and
now are) is dividing by the current wavelength PDFs any time we add to the
radiance sum for a pth.  Thus, vertices before the secondary wavelengths
are terminated aren't weighted incorrectly, by the post-termination PDFs...