Specify macros as compile flags with Visual Studio
CUDA integration in Visual Studio seems broken as even if "Use Host
Preprocessor Definitions" is checked, the host preprocessor definitions
are still not used when compiling device code.

To work around that, define the macros using --define-macro to avoid
CMake identifying those as macros and using the proper (but broken) way
of specifying them.