CMake: Ensure embedding is re-run after PTX generation
Having the embedding depend on the library used for compiling the CUDA
code down to PTX should be enough: according to CMake's documentation

> Additionally, if the target is an executable or library, a file-level
> dependency is created to cause the custom command to re-run whenever
> the target is recompiled.

However this was not happening, maybe due to it being an OBJECT library?
Or maybe some weird behaviour due to the modification timestamp on the
generated PTX file not being updated as the file gets re-generated.