GPU/wavefront integrator: don't hard-code HenyeyGreenstein as the only phase function.
Now a MultiWorkQueue templated on the phase function type is used for
enqueuing work items for sampling direct/indirect lighting in media.  With
Henyey-Greenstein as the only phase function currently available, this
reduces to the previous behavior, but it makes it easier to add new phase
functions in the future.

Also piped through time in a few places it was missing in the
medium-related work queues.