GPU/wavefront integrator: track per-ray depth independently of global wavefront depth.
This way, we can not increment the ray depth when it passes through a
surface that markes the boundary between different media, matching the
behavior of the regular CPU path.

In particular, this fixes some undesirable noise in the wavefront
integrator (and different behavior from the regular CPU path) that was due
to the ray depth being incorrectly tracked in scenes with participating
media and then Russian roulette inadvertently being applied after the very
first surface intersection.