Rework handling of allocators in GPU/wavefront integrator
The previous implementation repeatedly assumed that Allocators were
thread safe; this is not required for std::pmr allocators, but happened to
be the case when the default allocator is used on the CPU (since new/delete
are thread safe.) For the GPU, the CUDATrackedMemoryResource was
also thread-safe, but it also included some per-thread caching that doesn't
really belong in a memory resource.

Therefore, the WavefrontPathIntegrator (and friends) now no longer assume
that allocators are thread-safe, but instead explicitly use per-thread allocators
in parallel code. CUDATrackedMemoryResource is simplified to be just a
memory resource, and now we use a std::pmr::monotonic_buffer_resource
when we actually want chunked allocation.