Refactor ParsedScene
First here is now SceneStateManager, which implements the ParserTarget
interface and takes care of handling graphics state management. As parts
of the scene are defined, it passes them along to a SceneProcessor, a
new interface that currently has a single implementation that now takes the
name "ParsedScene".

In time, this will make it easier to try asynchronous creation of scene objects
as parsing occurs. For now, it seems to give a ~5% reduction in startup time
for the Moana Island scene.