Add MediumRayIterator to Medium interface
This is a much cleaner way for Medium implementations to directly specify
majorants for rays passing through them. In turn, they no longer need to
provide SampleT_maj() methods themselves; that can now be implemented in
terms of the iterators they return from their (new) SampleRay() methods.
This makes the general flow of medium sampling much cleaner, in addition to
making it easier to add new Medium implementations.

This comes with a 15-20% slowdown for scenes with media on the GPU (for no
obvious reason); that's worth chasing down at some point, but is worth the
code clarity for now.

This change also leads to a change in behavior for how scaled CuboidMediums
are treated: previously the SampleT_maj() method operated in medium space
whereas now it operates in rendering space, which leads to image
differences (for example, in the smoke-plume example scene. The current
behavior is arguably more sensible.

Also, renamed the existing Medium::Sample() method to SamplePoint() to
better distinguish between the two medium sampling methods now.