Windows: fix crash due to accessing managed memory at startup
Recent changes to parallelize more things during parsing/startup time turned out
to be crashing on Windows because we were simultaneously running GPU kernels
that accessed managed memory while also writing to (different) managed memory
on the CPU. That is, unfortunately, not allowed on Windows.

This fix uses the hammer of disabling the thread pool while the OptixAggregate
constructor runs.

Fixes #164