Fix bug with CPU writing to managed memory while OptiX is creating BVHs
On Windows, both CPU and GPU can't write to managed memory at the same time.
Previously, it was possible that the asynchronous creation of the sampler, camera,
film, etc., could still be running when OptiX BVH construction started. This was
especially likely with the Halton sampler + random digit permutations, which take
a while to initialize.

Fixes #246.